test: move WaitEvent class to test_common.h file
